Removed the - in several function names , renamed generics in RangeCell, and removed icon cell
