--- Frontend/src/Components/GeneralData/IconCell.js
@@ -1,7 +1,0 @@
-import { Box, Image} from "@chakra-ui/react";
-
-export default function IconCell(props) {
-    return(
-        <Flex>{props.boolean ? <Image boxSize={props.boxSize} src={props.image}/> : <Box h={props.boxSize}></Box>}</Flex>
-    );
-}
